updated project settings upon XCodes' request. Changed a log type to clear warning.

+++ Vehicles/Vehicle.m
@@ -35,7 +35,7 @@
         degrees = degrees % degreesInACircle;
     }
     
-    return [NSString stringWithFormat:@"Turn %d degrees.", degrees];
+    return [NSString stringWithFormat:@"Turn %ld degrees.", (long)degrees];
 }
 
 -(NSString *)changeGears:(NSString *)newGearName
